EasyCollege_Backend/requirements.txt
Flask
Flask-Cors
gunicorn
requests
beautifulsoup4
python-dotenv